PROJ1716: Initial work for automation of the API side from .proto
HU-3556 Implement automation of the API side of the interface

This patch is the initial work for client-side code generating from .proto.

While Brad is away, it looks good to put in to the project stream so other
developers can use it. Although it seems most of the implementation is complete,
any missing (updated) patch may be delivered when he is back if needed.

Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>

Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>