PROJ1716: Separate out the ATL autogenerated code.
HU-530: Split out autogenerated code

All the ATL autogenerated code was being added to the pbc header
and c files. This produced big ugly files, but also caused problems
because it required the client to have a dummy impl function for
compilation to work. It also meant that both the client and server
had extra code they didn't need.

This patch splits out the ATL code into the following files:
<proto name>_types_auto.h <-- contains the ATL generated types
<proto name>_api_auto.[ch] <-- contains ATL generated client code
<proto name>_impl_auto.[ch] <-- contains ATL generated server code

The temporary fix to generate a weak impl function for the client is also
reverted as part of this patch.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>