PROJ1716: Some parameters were missing from api and impl functions
HU-561: Fix bugs in atl generated code

Repeated fields need to be passed with a parameter for the type, as well
as a size_t to indicate the number of repeated fields.
The size_t parameter was missing for repeated messages (although it
was there for repeated primitive types).

This is now fixed.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>