PROJ1716: Fixed api and server autogenerated memory bugs
HU-561: Fix bugs in atl generated code

Some mallocs were too small because they were using the size
of the ATL struct and not the PBC struct.
Also, the PBC structs were not being initialised properly
before transmission.

Finally, the server side was not allocating memory for the parameters
to be sent to the impl function, which was causing a segmentation fault.

This is now fixed.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>