PROJ1716: Fixed "string" copy in auto-code
HU-515 As a developer I need a single consistent way to build and link to the auto-generated code

When copy a string by using "strncpy(dst, src, strlen(src))", the trailing '\0'
may not be copied to the destination string. This has been changed to use
"strcpy(dst, src)" instead.
Also in _api and _impl functions, fixed an incorrect use of malloc and then
just overwriting with a new pointer which causes memory leak as well.

   src = malloc();
   src = dst;

And tidied up some code in auto-code.

Reviewed-by: Florian Roeske <florian.roeske@alliedtelesis.co.nz>