PROJ1716: Retain package/service/message/rpc names as it is in auto-code
HU-554: Retain lower-/upper-case as it is in the auto-generated code

Original protobuf-c compiler converts names defined in .proto to Camel-case
for struct and lower-case for members, etc.
This is going to be inconvinient for users to match with .proto and the auto-code.

Instead of changing many placing converting names, simply modified help
functions, which is simpler and efficient to handle the case of converting
"." to "_" properly as well.

Reviewed-by: Florian Roeske <florian.roeske@alliedtelesis.co.nz>