PROJ1716: Handle lost subscriber connections
HU-352: As a process I need to inform others when something happens

When a subscriber server is not reachable by a publisher client.
The subscriber server entry will be deleted in the subscriber list
on the publisher. When a subscriber restarts and registers again
the subscribtion will be readded to the subscriber list on the
publisher.

Reviewed-by: Stan Xiang <stan.xiang@alliedtelesis.co.nz>

Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>