PROJ1716: Fix type of rpc client in autogenerated code
The type of the rpc client being passed to the api function was different
to the type returned from the create client function.
This required the developer to cast the client in the call to the api.

This patch changes the type in the api paremeter list to be the same as the
developer is using, and then casts inside the autogenerated code to the
correct type.

Reviewed-by: Florian Roeske <florian.roeske@alliedtelesis.co.nz>

Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>