PROJ1716: copying of returned values to developer space was not reliable.
HU-615: rpc results are not being copied properly in autogenerated api functions

The generated closure function was never updating the value of the "has_..."
field for optional fields.
The generated api function checks the "has_" value before copying values to
the developers out parameters.
As a result, it was random whether or not the developer output parms had their
values set to the returned (from rpc) values.

Fixed this by changing the generator to update the "has_" fields in the closure
function. Message struct pointers are now also checked before attempting to
use them.
msgR (in the api code) was also initialised to stop random values in the "has_"
fields and pointer values.

Reviewed-by: Florian Roeske <florian.roeske@alliedtelesis.co.nz>

Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>