PROJ1716: generalise socket addr for other addr options
JIRA HU-597

To use user defined transport it needs to know where it is sending the message
which will be dependent upon the user defined structure that is set within the
user defined transport structure.  This is a different type of connection info
to the socketaddr that is used for other transport types and so I have refactored
the socketaddr to allow the either or use of it and the user defined pointer.

Reviewed-by: Florian Roeske <florian.roeske@alliedtelesis.co.nz>

Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>

Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>