PROJ1716: Fix problem in previous patch
The last patch changed the way bytes fields were freed. Unfortunately,
this relied on another (unfinished) patch which changed how bytes
fields were defined in the C code.
As a result, the generated code now doesn't compile.

The temporary fix is to revert the problem line until the other
patch is ready.

Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>

Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>