PROJ1716: Fix handling of bytes fields
HU-620: Fix handling of bytes fields

A patch went in earlier that changed the generated type for bytes fields.
This is causing some problems in the lower level rpc code that still assumes
the type is unchanged (and is still ProtobufCBinaryData).
Fix this by changing some of the generated code back to using the
ProtobufCBinaryData type.

Note: This is tricky because the ProtobufCBinaryData field is a structure
but gets created as a struct and not a pointer to a struct (like all other
structs in the generated code). This means handling the bytes field is like
a message in some ways and like a primitive in others.
As a result, this patch isn't quite complete, but covers all the cases we are
currently using.
One case is currently not handled - returning a bytes field by itself to the
developer through the API. However, returning a bytes field that is part of
another message is fine.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>

Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>