PROJ1716: Change CPG server & client to share the connection
The CPG connection is meant to be shared between the client & server.  There
is a lot of wasted time & memory if the connection is instantiated when
sending every message.

This has also involved changing the names of functions to better reflect
their usage and the addition of function comments.

Also updated the test client for cpg.

Reviewed-by: Florian Roeske <florian.roeske@alliedtelesis.co.nz>

Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>