PROJ1716: Implement shorter integer fields
The smallest integer supported by the google compiler was 32 bits.
When replacing existing messages with autogenerated ones, they can become
quite a bit bigger if the original used int8 and int16 fields.

The base protobuf compiler has now been extended to provide support for
uint8, int8, uint16, and int16 type keywords. This patch updates the
protobuf-c compiler to allow use of these new types. It also updates
the generator to produce the correct autogenerated types and code for the
new smaller types.

Finally, the packing and unpacking in cmsg has been updated to handle these
types as int32 or uint32 (for signed and unsigned types respectively) when
going on and off the wire.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>

Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>