PROJ1716: Use sockadddr_tipc instead of generic structure
Converts the use of the generic addressing structure to the tipc addressing
structure for the tipc broadcast listen function.

The use of the generic addressing structure have been left in the standard
tipc file, as to change these would require functional changes. It would be
nice to get these changed over at some time in the future.

Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>

Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>