PROJ1716: Change the way enum names are generated
The way enum names were generated caused some problems for developers.
Specifically, the final name in the C code was:
<pkg name>_<enum type>_<name defined in .proto>
Part of the problem is that the <name defined in .proto> has to be unique
even though the resultant C code name is different so that two values with
same name (but in different <enum types>) would be rejected by the compiler.

This patch changes the way the final name of the enum value is generated to be:
<pkg name>_<name defined in .proto>

This will allow the developer to create a unique <name defined in .proto>
without the resultant C code being too long and silly.

Note: this patch also updates the cmsg test program to get it working again
after recent changes, and to add an enum to the test message sent.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>