PROJ1716: Remove double mutex lock
cmsg_pub_invoke () held the mutex on subscriber_list, and then sometimes called
cmsg_pub_subscriber_remove (). cmsg_pub_subscriber_remove () also attempts to
acquire the mutex to the subscriber_list. Therefore the thread would lock
waiting for itself to release the lock.

This patch creates a thread unsafe version of cmsg_pub_subscriber_remove ()
that can be used if you already hold subscriber_list_mutex.

It also moves the call to get the next subscriber in the list in front of the
call to remove the current subscriber from the list, as this could cause
problems.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>

Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>