PROJ1716: Fix autogenerated header file ifndef/define
The "#ifndef/#define" at the start of every ATL generated header file could
produce compile warnings if the .proto file name contained a hyphen "-".
The hyphen would end up being included in the define which made the compiler
determine that the define finished at the hyphen. This produced a warning that
the define had extra tokens, and was missing whitespace at the end of the
define.

This patch simply filters the name used in the ifndef/define to convert
hyphens to underscores, and convert other non-legal chars to the hex code
for that char.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>