PROJ1716: Add support for sending responses when queuing RPC calls
When the calls are queued or dropped by the server the client still needs to
be sent a response to unblock it.

The response has a special status code and has no data.  This at least allows
the client to move on.

If the function is expected to have some return values, they will not be set
at all.  The Applicaiton is expected to make sure they do not queue or drop any
methods that get values back (non return codes).

Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>