PROJ1716: Ensure output repeated fields are big enough before copying (autogen)
HU-888: Ensure output repeated fields are big enough before copying (autogen)

The autogenerated code in the API wasn't checking the size of memory allocated
by the caller. Instead it was just copying over whatever it received from the
call to the RPC.

This patch fixes this by adding a check of the allocated size of the
destination, and only copying what will fit if the source is bigger than the
destination.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>