PROJ1716: Keep RPC connection open unless errored
Everytime an RPC occurred, the client would connect to the server, send
the message, and then close the connection.
When an interface had a lot of RPC calls to make, the constant opening
and closing of the connection was impacting performance quite badly.

This patch changes things on the client and server side so that the
connection is kept open unless an error occurs. This way the connection
only needs to be opened if the last RPC failed for some reason.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>