PROJ1716: Reuse the subscribed client inside the publisher
The cmsg publisher created a client connection and then destroyed it
for every notification it sent.
This can cause a deadlock situation between the subscriber and the publisher.

This patch fixes this by adding the client to each subscription so that the
client will only be created on first notification sent to that subscriber,
or when the connection is lost.

Note: This results in a copy of the client being created for each
subscription, which is a bit inefficient if the same subsriber has
multiple subscriptions.
This was done deliberately as the alternative was to have a single client in
another table somewhere, that pointed back to the transport which was used as
the subscription entry. The problem with that is if a subscriber only
cancelled one subscription, and that was the transport entry that the client
was using, the client would then have to walk the list to try and find another
transport entry to point to and use.
This makes things horribly complicated, especially in the case of determining
when a subscriber has no more subscriptions, so it was decided to use the
simpler solution.

Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>