PROJ1716: Allow applications to initiate pub->sub client connections
This patch adds a couple of functions to allow an application to initiate the
connection process from the publisher to the subscribers.
The first function will create clients for every subscription entry.
The second function will only create clients for a particular subscriber
(where the transport is used to identify all subscriptions from that
subscriber).

This functionality can be helpful to prevent deadlocks by allowing an
application to only initiate connections if/once it knows the subscriber is
ready to receive it.

As this patch builds on the previous patch, I also fixed the formatting
of some code added in the previous patch.

Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>