PROJ1716: Change to start the connection in the client initiate functions
Previously when the client was created, the connection wasn't being
established until the first send. The initiate subscriber functions were added
to allow the application to decide when to initiate the connections.
Unfortunately, these functions only created the client, but didn't start the
connection.

This patch changes the initiate subscriber connections functions to fulfil the
original intent (ie to also start the connection once a client is created).

Also added some locking around the use of the subscriber list in the initiate
subscriber connections functions.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>