PROJ1716: Fix NULL client crash in publisher
The client belonging to a subscription entry is only created if the
application asks for it (by calling initiate_all_subscriber_connections)
or upon the first notification to be sent to that subscription.
If the subscription is cancelled before the client is created, a NULL
pointer was being followed (and that way leads to madness).

This patch fixes this by testing the client pointer before attempting to
destroy it.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>