PROJ1716: Do not initiate connection on publisher when queuing is enabled
When queuing is enabled on a publisher, initiating client connection causes
problem that it may block until the subscriber accepts or timeout happens.
But the messages in the queue is designed to be processed normally by different
function call (or even from different thread in poehw.)

So initiating client connection is done only when queuing is not enabled.

Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>

Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>