PROJ1716: Modify CMSG header to allow for other types of internal messages
This is adding the ground work for healthcheck support.

The CMSG headre has been changed to be 20 bytes long of the form:
msg_type
header_length
message_length
method_index
status_code

Not all fields ae applicable to all msg_type's however it was easier to implement
it this way at this stage.  The fields are better documented in protobuf-c-cmsg.h

Along the way of debugging some issues I have changed a number of debug error
logs over to the new way of logging errors so they will be visible.

Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>