PROJ1716: Add echo support to the client & server
ECHO support allows a client to send an echo request to a server and to get a
reply as soon as the server processes the request.

This feature can be used to determine if the server is working - therefore for
healthcheck purposes.

There are expectations that the client is not used for other messages and that
the transport used is RPC (twoway).

Current echo reply reception relies upon the application listening to the
clients socket and calling a client fn to receive the echo reply when data is
on the socket.

Further development could be done to make it block if required but this isn't
needed at present.

Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>

Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>