PROJ1716: Shutdown cmsg client connection when destroying client
Previously, the owner of the client was expected to shutdown the connection
before destroying the client, but only one place did this.
Part of the problem is that it was expected that the destroy would do a tidy
shutdown of the client, and not leave the connections open.
The lack of shutting down the connection led to file descriptor leaks.

This patch fixes the problem by adding the connection shutdown to the client
destroy.
To ensure that connections aren't attempted to be closed twice, the transport
client close functions have been updated to set the socket to -1 after
shutdown, and to check that the socket is not -1 before attempting to close
it.
Also, client_new is updated to initialise the socket to -1.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>