PROJ1716: Fix "visible" autogenerated code to meet coding standards
The developer facing autogenerated code didn't comply with the atlas coding
standard of putting a space between the function name and the bracket that
starts the parameter list.
This made more work for the developer as they couldn't just cut and paste the
api and Send function declarations - they had to add the space themselves.

This patch adds the space in.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>