PROJ1716: Lower server connect error message to debug
An error message was being generated every time the server lost its
connection, but this is a regular normal condition (say when a stack member
leaves the stacket).
This was causing lots of scary errors in the system log for perfectly normal
events.

The fix is to downgrade the error to be debug - it can be switched on if the
developer is tracing a problem, but the normal case will be for no log
message.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>