PROJ1716: Fix memory allocation size of repeated primitives (autogen)
In the autogenerated message copy code, repeated primitives were having the
wrong amount of memory allocated (enough memory for an array of pointers, not
for an array of <types>).
This is usually harmless as it just results in either the right amount of
memory anyway, or too much memory, depending on the type.
However in the case of 64 bit integers, not enough memory will be allocated
which is bad.

This patch corrects the size of the memory allocated.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>