PROJ1716: Fix the type used to send enum in autogen code
The autogenerated code in the API was treating enums as uint8 instead of
uint32 (which is what is sent through the lower layers).
This worked for individual enums, but resulted in not enough memory being
allocated for repeated enums. This caused a crash.

The fix is to change the type used in the autogenerated code from uint8 to
uint32.

Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>

Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>