PROJ1716: Changed "_service" name in impl definition to just "service"
The autogenerated definition of the impl functions used the name "_service" as
the first parameter.
Most developers changed this to just "service" when implementing the impl, but
a rule that we want to enforce is that the parameter names in the impl
implementation must match the impl definition.

This patch changes the autogenerated to just use "service" so that we can
enforce the rule that definition and implementation must match (without needing
exceptions to the rule).

Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>

Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>