CR00040625: Fixed openais timeouts in standalone x908 when using XEM12T
Description:
On a standalone x908 system with XEM12Ts, EXFX was failing to initialise
connectivity to CPG.  When in standalone, EXFX tries to immediately
connect to CPG and has a 10 second timeout.  This timeout was arbitrary
and didn't exist with the obsolete IPC code.  The reason a timeout was
originally chosen was to reconcile differences between Fastfailover and
EXFX.
The solution is to increase the timeout.
30s has been chosen.

Testing:
Tested on x908 with 20 consecutive reboots.  Before fix error was seen
18/20 times.  With fix 0/20 times.
Tested on x908 stack with reboot and fastfailover.  No issues seen.

Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>