CR00040921: Use a reference to queue_mutex instead of a copy
A copy of a mutex has an undefined behavior (see the manpage for
pthread_mutex_init), and should not be used. Instead, a mutex should be
referenced with a pointer to the original.

This patch corrects an issue in cmsg_send_queue_process_all which resulted
in locks being held and never released, as well as other places where the
use of mutex copies was seen.

Reviewed-by: Florian Roeske <florian.roeske@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>