CR00040992: Increase CPG connection timeout.
On a SBx8100 chassis with SBx81CFC400 controller EXFX openais manager
timed out to connect to corosync on bootup.

The root cause is that the openais manager tried to connect before
Corosync has started. The timeout value selected is 30 seconds which
is arbitrary.

So increasing the timeout to 180 seconds to guarantee that
exfx does not fail to initialize a connection with CPG.

Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>