CR00041118: CMSG error "response message not valid or empty"
When ecofriendly button is pushed, the following error log is generated.

  ERR(CMSG):cmsg_client_invoke_rpc 291: [CLIENT] error: response message not valid or empty

This is only error log and actual ecofriendly LED is properly set to on/off.

This error happened because the server (hostd) sent back a CMSG response with
empty message body which is designed that way. But CMSG API API incorrectly
judged message was arrived in bad format.

Now this has been fixed by properly checking the return code and allowing
empty message body.

Also updated the log message to display the method causing this error for
better debugging.

Reviewed-by: Florian Roeske <florian.roeske@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>