CR00041118: Lower log level to DEBUG for CMSG log "Connection refused"
It appears that the following error log can happen any time when a LIF or
a stack node leaves while a CMSG client is trying to connect or send.

ERR(CMSG):cmsg_transport_tipc_connect 44:
  [TRANSPORT] error connecting to remote host (port 9700 inst 4): Connection refused

As this is mostly expected behaviour and it may unnecessarily alert customers.
Also EasyIPC did not report this error before, so it would be ok to make this
change in CMSG to be debug-level message. And if needed for debugging those
messages are visible by "terminal monitor" or changing log level in IMISH.

* cmsg_transport_tcp_connect(),
  cmsg_transport_tipc_connect(), etc:
  Changed to return a negative integer on error (mostly it's "-errno")

* cmsg_client_invoke_rpc(),
  cmsg_client_invoke_oneway(),
  cmsg_client_send_echo_request():
  Changed the log level from ERROR to DEBUG for "connect" failure

* Also added the error code to debug log messages

Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>