PROJ1756: Timing analysis tools
ISSU-71: timing analysis for different message sizes

Added cmsg-prof tool for analysing cmsg performance on the
switch. This is a basic test program which creates a tipc
server and/or a client and logs the taken call time to csv file.

Building of cmsg-prof has to me enabled manually with the
--enable-cmsg_prof and HAVE_CMSG_PROFILING option.

The local-install script has been changed accordingly.
New options for the local install script.
  -p|--enable-cmsg-prof
  -v|--enable-vcstack
  -h|--help

Change-Id: I8354be604dce542bfc9374c0bcd0b9ec8c844e7a
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>