PROJ1756: Remove closure function from client side
ISSU-164: Remove use of "closure_data_function" from the api

The use of the closure function was something inherited from protobuf-c and
only made sense with that queuing/dispatch model.
That model has since been replace with cmsg, so the closure function is now
unnecessary. It is also inefficient as it forces an extra copy of the recv
message along with the processing of the mallocs and copies required for the
deep copy.

This patch removes the need for the closure function by getting the received
message up from the invoke and copying directly from it.
A client closure data structure was created to get the recv message pointer
and the memory allocator back from the invoke, without having to modify the
invoke function. Modifying the invoke function would force changes to many
other parts of protobuf-c/cmsg.
Because the server already defined a closure data structure, this has been
renamed to cmsg_server_closure_data.
The client closure data has been called cmsg_client_closure_data.

Also, this patch splits out the freeing of unknown fields logic from the
free_unpacked function into its own function (which free_unpacked now calls).
This is to allow the API (in a later patch) to just free the unknown fields
from a message before passing it on to the developer.

Change-Id: Ideec342cf8bc6b1b45c80d1e6de22ffa92f9f9a4
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Florian Roeske <florian.roeske@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>