PROJ1756: Change to get cmsg error codes back to the api caller.
ISSU-113: As a developer I need to get any errors from failed CMSG API calls.

The API currently can't return meaningful error codes to the developer because
the lower level functions it calls doesn't return anything.

This patch changes the lower level functions to return a status to the API so
the API can return it to the developer.

Note: This change required changing the autogenerator, which then required the
subscriber service code to be re-generated. Because the saved autogenerated
code came from a much older version of the autogenerator, the code in cmsg
that called it also had to be updated.

Change-Id: I13b417c3770e569d62d6cb5410147482499c635b
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Florian Roeske <florian.roeske@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>