PROJ1756: Change the api to have the pbc structs passed directly.
ISSU-165: Convert autogenerated API/IMPL to take PBC structures as arguments

To make things more efficient (ie preventing multiple malloc's and msg
copying) the developer will now create the message themselves, and pass it to
the API. The impl will also receive the msg structure directly.
This is rather than the autogenerator splitting out the high level fields as
parameters, and the copying and assembling/dissasembling the msg structures
under the hood.

To make the transition easier, the generator continues to generate the old API
as well. The new generated api functions will have the name "...apiNEW...".
Once the conversions to the new api are complete, the old api functions will
stop being generated, and the new api functions will be generated with the old
names.

Note: This patch only creates the new API functions. It doesn't change
anything on the server side (impl and server Send functions) - this will be
done in the next patch.

Change-Id: If2b2689cb56780f867f98df6bb9cf89f21e79190
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Florian Roeske <florian.roeske@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>