PROJ1756: Change the impl to have the pbc structs passed directly.
ISSU-165: Convert autogenerated API/IMPL to take PBC structures as arguments

To make things more efficient (ie preventing multiple malloc's and msg
copying) the developer will now create the message themselves, and pass it to
the api. The impl will also receive the msg structure directly.
This is rather than the autogenerator splitting out the high level fields as
parameters, and the copying and assembling/dissasembling the msg structures
under the hood.

To make the transition easier, the generator continues to generate the old
impl as well. The new generated impl functions will have the name
"...implNEW...".
Once the conversions to the new impl are complete, the old impl functions will
stop being generated, and the new impl functions will be generated with the old
names.
Also, impl stubs will be generated in a temporary file that is not built.
The developer can cut and paste these into the source files to allow builds to
commence, and gradually convert over the old impls to the new stubs.

Note: This patch only creates the new IMPL and server send functions. It
doesn't change anything on the client side (api) as this was done in an
earlier patch.

Change-Id: Idefb29544c1d856f947bfc7d51f5858ea9f46f8c
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: Florian Roeske <florian.roeske@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>