PROJ1756: Remove method hash table look up to get the method index
There is a protobuf-c function which returns the descriptor by checking the method name.
This function is never used and made a slight variation to the function to return the
method index. This function is used on the receiver side to get the proper method_index.

This patch removes all the method hash table implementations.

Change-Id: I95d9289fbe93e289527b5a08d74bf3dacbe01bb5
Reviewed-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>