PROJ1756: Turn off the autogeneration of the old impl and send.
This also renames the new impls and send functions to be the same name as the
old ones.
Finally, the cmsg test program is updated to remove the use of the old impls
and sends.

Note: Turning off autogeneration of the old API will be done in the next
patch.

Change-Id: I5a98bfe91686b1aad841fa2a2067c69b3b194651
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>