PROJ1756: Change CMSG client to return queued or dropped to notification APIs
Notification APIs are one way messaging technique. CMSG library queues or drops
the notifcation according to the action required. But wrong error code is returned
in these cases. This patch allows CMSG to return CMSG_RET_QUEUED or CMSG_RET_DROPPED
according to the action taken place in the library.

Change-Id: Id296110a3245a6c3be828b6c95dd84b16ee77ccd
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>

Signed-off-by: John Thompson <john.thompson@alliedtelesis.co.nz>