CR00042111: Make QUEUED and DROPPED available to the client when they occur
When the server queued or dropped an RPC this was not being returned to the API
fn.  When it was happening the out values of the RPC do not get set and as result
when the value was initialised to error then the caller of the API was returning
error.  Previously under EasyIPC the server sent back 0 (signifying OK) when the
RPC was queued or dropped and so the API fn would return this value.

To resolve the problem we make use of the invoke_return_state field in the client
to return the return value of the RPC to the caller of API.  It wasn't being set
for QUEUED or DROPPED and so this patch updates it for server side queuing.

Change-Id: Id16821e6f24a73f2f5928f7815672c552f2345d5
Reviewed-by: Scott Bamford <scott.bamford@alliedtelesis.co.nz>
Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>