PROJ1756: Respond to unknown method
Jira: ISSU-12

Currently, when a CMSG server receives an unknown method, it asserts. It needs
to be able to appropriately handle and respond to an unknown method being
received, as this is a likely to happen in ISSU conditions.

Therefore, if cmsg_tlv_header_process receives an unknown method string, it will
return CMSG_RET_METHOD_NOT_FOUND. This in turn triggers the server to reply to
the client with an empty message with a status code of
CMSG_STATUS_CODE_SERVER_METHOD_NOT_FOUND. When a client receives this message,
it returns CMSG_RET_METHOD_NOT_FOUND to the api caller.

Change-Id: I8a4e145e35e2a1f75ab3a53c16daf9753ded8965
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>