PROJ1756: Unlock subscriber list while sending notifications
Jira: ISSU-376

On start-up, sometimes a CFC in a VCS+ stack is unsubscribed from port event
notifications from exfx. This means it no longer is informed if a port goes up
or down. This was happening because of a deadlock caused by the notification
publishers subscriber_list_mutex.

Now, in the publisher invoke function, the publisher releases the mutex on the
subscriber list while it is attempting to send the notification. It only grabs
the mutex when it needs to know the next subscriber to send to.

In order to stop the subscriber from being removed from the list while it is
being used to send a notification, the removal process has been changed. Now,
when a subscriber needs to be removed from the list, it is marked with a
timestamp. The next time that the list is iterated by the invoke function,
it notices any subscribers that have been marked for removal, and calls
cmsg_pub_subscriber_remove_expired_entries to remove them. Entries are removed
from this list if there attempted_remove_time is more than (5 *
CMSG_TRANSPORT_CLIENT_SEND_TRIES * CMSG_TRANSPORT_TIPC_PUB_CONNECT_TIMEOUT /
1000 ) seconds old. This means that if someone attempts to remove a subscriber
currently being used, the memory won't be freed out from under them.

Invoke will also check to see if the subscriber has been marked from removal,
and won't attempt to send using a subscriber that has been marked.

Change-Id: I925b54c73f0ac856b4d86e554bcd2dce0b3c60c2
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Brad Carthew <brad.carthew@alliedtelesis.co.nz>